requirements.txt
Flask
flask-restful
psycopg
vk-api==11.9.9
python-dotenv





